Switch to Resource for handling, well... resources
Main reason is because using "src/main/resources" on projects with Maven and running them in the root directory really gives an allusion that I'm referencing a relative path.
When you compile something into a .jar, resources get moved to the root of the jar and the way they're referenced is way different. This aims to fix that: by changing `new File("src/main/resources/file.png")` to `new Resource("file.png")`.